requirements.txt
pillow=8.0
numpy
qrtools
pyzbar
qrcode
pyqrcode
tkinter